Time: 0 ms (100%), Space: 9.4 MB (6.15%) - LeetHub

+++ 3754-concatenate-non-zero-digits-and-multiply-by-sum-i/3754-concatenate-non-zero-digits-and-multiply-by-sum-i.cpp
@@ -10,7 +10,7 @@ public:
                 sum += c - '0';
             }
         }
-        long long x = stoi(t);
+        long long x = stoll(t);
         long long res = x * sum;
         return res;
     }
